@@ -13,60 +13,61 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.2.11",
-    "@angular/cdk": "^7.3.6",
-    "@angular/common": "^7.2.11",
-    "@angular/compiler": "^7.2.11",
-    "@angular/core": "^7.2.11",
-    "@angular/flex-layout": "^7.0.0-beta.24",
-    "@angular/forms": "^7.2.11",
-    "@angular/http": "^7.2.11",
-    "@angular/material": "^7.3.6",
-    "@angular/platform-browser": "^7.2.11",
-    "@angular/platform-browser-dynamic": "^7.2.11",
-    "@angular/router": "^7.2.11",
-    "@ng-bootstrap/ng-bootstrap": "^4.2.1",
-    "@types/underscore": "^1.9.2",
-    "angular2-moment": "^1.9.0",
-    "bootstrap": "^4.5.2",
-    "core-js": "^2.6.0",
-    "jquery": "^3.5.1",
-    "moment": "^2.24.0",
-    "ng2-pdf-viewer": "^5.3.4",
-    "ngx-file-drop": "^8.0.7",
-    "ngx-logger": "^3.3.12",
-    "ngx-socket-io": "^2.1.1",
-    "popper": "^1.0.1",
-    "popper.js": "^1.12.9",
-    "rxjs": "^6.4.0",
-    "tslib": "^1.9.0",
-    "typeface-open-sans": "^1.1.13",
-    "underscore": "^1.9.1",
-    "zone.js": "^0.8.26"
+    "@angular-devkit/build-webpack": "0.1002.0",
+    "@angular/animations": "10.2.0",
+    "@angular/cdk": "10.2.6",
+    "@angular/common": "10.2.0",
+    "@angular/core": "10.2.0",
+    "@angular/flex-layout": "10.0.0-beta.32",
+    "@angular/forms": "10.2.0",
+    "@angular/material": "10.2.6",
+    "@angular/platform-browser": "10.2.0",
+    "@angular/platform-browser-dynamic": "10.2.0",
+    "@angular/router": "10.2.0",
+    "@ng-bootstrap/ng-bootstrap": "4.2.1",
+    "@types/underscore": "1.9.2",
+    "bootstrap": "4.5.2",
+    "core-js": "3.6.5",
+    "jquery": "3.5.1",
+    "moment": "2.24.0",
+    "ngx-extended-pdf-viewer": "5.3.0",
+    "ngx-file-drop": "8.0.7",
+    "ngx-logger": "3.3.12",
+    "ngx-moment": "5.0.0",
+    "ngx-socket-io": "3.2.0",
+    "popper": "1.0.1",
+    "popper.js": "1.12.9",
+    "rxjs": "6.6.3",
+    "tslib": "1.9.0",
+    "typeface-open-sans": "1.1.13",
+    "underscore": "1.9.1",
+    "zone.js": "0.11.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.10.0",
-    "@angular/cli": "7.3.7",
-    "@angular/compiler-cli": "^7.2.11",
-    "@angular/language-service": "^7.2.11",
-    "@types/jasmine": "^3.3.12",
-    "@types/jasminewd2": "~2.0.2",
-    "@types/node": "^11.12.2",
+    "@angular-devkit/build-angular": "0.1002.0",
+    "@angular/cli": "10.2.0",
+    "@angular/compiler": "10.1.1",
+    "@angular/compiler-cli": "10.1.1",
+    "@angular/language-service": "10.2.0",
+    "@types/angular-material": "1.1.72",
+    "@types/jasmine": "3.3.12",
+    "@types/jasminewd2": "2.0.2",
+    "@types/node": "11.12.2",
     "@types/webrtc": "0.0.24",
-    "codelyzer": "^5.0.0",
-    "jasmine-core": "^3.3.0",
-    "jasmine-spec-reporter": "^4.2.1",
-    "karma": "^4.0.1",
-    "karma-chrome-launcher": "^2.2.0",
-    "karma-cli": "^2.0.0",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
-    "karma-jasmine": "^2.0.1",
-    "karma-jasmine-html-reporter": "^1.4.0",
-    "protractor": "^5.4.2",
-    "rxjs-tslint": "^0.1.8",
-    "ts-node": "~4.1.0",
-    "tslint": "^5.14.0",
-    "typescript": "^3.2.4",
-    "webpack": "^4.29.6"
+    "codelyzer": "5.0.0",
+    "jasmine-core": "3.3.0",
+    "jasmine-spec-reporter": "4.2.1",
+    "karma": "5.2.3",
+    "karma-chrome-launcher": "2.2.0",
+    "karma-cli": "2.0.0",
+    "karma-coverage-istanbul-reporter": "1.2.1",
+    "karma-jasmine": "2.0.1",
+    "karma-jasmine-html-reporter": "1.4.0",
+    "protractor": "7.0.0",
+    "rxjs-tslint": "0.1.8",
+    "ts-node": "4.1.0",
+    "tslint": "5.14.0",
+    "typescript": "3.9.2",
+    "webpack": "4.40.0"
   }
 }
