@@ -1,31 +1,32 @@
-import { Component, OnInit, Inject } from '@angular/core';
-import {
-  MAT_DIALOG_DATA
-} from "@angular/material";
-import { NGXLogger } from 'ngx-logger';
-import { MatDialogRef } from '@angular/material';
+import { Component, OnInit, Inject } from "@angular/core";
+import { MAT_DIALOG_DATA, MatDialogRef } from "@angular/material/dialog";
+import { NGXLogger } from "ngx-logger";
 
 @Component({
-  selector: 'app-preview-modal',
-  templateUrl: './preview-modal.component.html',
-  styleUrls: ['./preview-modal.component.scss']
+  selector: "app-preview-modal",
+  templateUrl: "./preview-modal.component.html",
+  styleUrls: ["./preview-modal.component.scss"],
 })
 export class PreviewModalComponent implements OnInit {
-
   hasDialogAction: Boolean = true;
 
   pdfSrc: Object;
 
-  constructor(@Inject(MAT_DIALOG_DATA) public data: any, private logger: NGXLogger, private dialogRef: MatDialogRef<PreviewModalComponent>) { 
+  constructor(
+    @Inject(MAT_DIALOG_DATA) public data: any,
+    private logger: NGXLogger,
+    private dialogRef: MatDialogRef<PreviewModalComponent>
+  ) {
     dialogRef.disableClose = true;
-    this.hasDialogAction = (data.hasDialogAction != undefined) ? data.hasDialogAction : this.hasDialogAction;
+    this.hasDialogAction =
+      data.hasDialogAction != undefined
+        ? data.hasDialogAction
+        : this.hasDialogAction;
   }
 
-  ngOnInit() {
-  }
+  ngOnInit() {}
 
-  closeModal(){
+  closeModal() {
     this.dialogRef.close();
   }
 }
-
